렌더링 분리
